Add C implementation for day 1; use fixed-len array in Rust impl
As a (likely, fastest) base benchmark, I figured it would be a good
idea to implement a solution to day 1 in C.  In this implementation, I
used a fixed-length array (I don't even know if C has a
variable-length array/vector).  I thought it might be nice to
implement that in Rust as well, so that is what I did.