Clean up AoC Day 7 2022 Julia solution
I made the mistake of writing a recursive solution to fully parse and
model the file system over just keeping track of the directory paths
in a dictionary, and their total sizes, when I have not got much
experience with recursion, nor trees, and I am very stupid.  As such,
I took a good couple of hours to figure it out, and took a lot of
debugging.  This commit cleans up that horrific code.  (Don't get me
wrong, it's still horrific, but just slightly more presentable.)
This, coupled with the fact that I could not start the problem till
around 5 hours after it came out, meant that I did _not_ do well this
day on the leaderboard!