Solve AoC Day 13 2022 in Julia
I had a friend's birthday dinner tonight, and we went to see a show
afterwards, so I could only start this 5 and a half hours after the
problem came out.  It's a shame, as I really enjoyed this one, and
found it quite simple (utilising Julia's multiple dispatch and
customisable sorting); it only took less than 45 minutes to complete
both parts.