Initial work on AoC Day 17 2022 in Julia
I have a really strange bug and I'm not quite sure how to fix it, as
it all looks to be working perfectly, except the test case does not
have the right number.  Nevertheless, I am committing the current
state of code, to show I have been working on it somewhat.