Add summary of solutions to early AoC days of 2022
After the 13th, I realised that it was annoying when I was looking
back at old AoC solutions but couldn't for the life of me remember
what any of the code did or why, and I didn't want to have to load up
the AoC website for that day of the previous years and read the
description there to find out what it was.  Hence, I started writing
short summaries of the problems, and a quick description of my
solution, and any other worthwhile notes.  As I only started doing
this after the 13th, I retrospectively add these notes/summaries in
this commit.